fix: coderabbit issues
