+++ packages/server/envoy-wasm-ts.yaml
@@ -55,5 +55,5 @@ static_resources:
 admin:
   address:
     socket_address:
-      address: 0.0.0.0
+      address: 127.0.0.1
       port_value: 9903
